Move AppBar and Drawer into widgets folder. Add backgroudColor to Drawer.kt
