+++ app/src/main/java/com/example/smart_job_finder_v2/ui/screens/Likes.kt
@@ -9,7 +9,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
-import com.example.smart_job_finder_v2.ToolBar
+import com.example.smart_job_finder_v2.ui.widgets.ToolBar
 import com.example.smart_job_finder_v2.ui.widgets.BottomBar
